Move discover_packages() method to outside config.py
config.py should handle only parsing command line and config YAML file,
nothing else.
