+++ lib/config.py
@@ -14,7 +14,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import argparse
-import logging
 import os
 import sys
 
@@ -22,9 +21,7 @@ import yaml
 
 from lib import log_helper
 from lib import utils
-from lib.constants import REPOSITORIES_DIR
 
-LOG = logging.getLogger(__name__)
 LOG_FILE_NAME = 'builds.log'
 BUILD_REPO_ARGS = {
     ('--packages-metadata-repo-url',):
@@ -116,48 +113,6 @@ def get_config():
         config_parser = ConfigParser()
         config_parser.parse()
     return config_parser
-
-
-def discover_packages():
-    """
-    Simple mechanism for discoverability of the packages we build.
-
-    A discoverable package, and thus potentially buildable, will be assumed as
-    any directory name under the packages metadata git repository directory containing
-    a yaml file with the same name.
-    Considering the example:
-
-    versions
-    +-- kernel
-    |   +-- kernel.yaml
-    +-- libvirt
-    |   +-- libvirt.yaml
-    |   +-- someother_file_or_directory
-    +-- not-a-package
-    |   +-- not-following-standards.yaml
-    +-- file
-
-    "kernel" and "libvirt" will be discovered, "not-a-package" and "file"
-    will not.
-    """
-    config = get_config().CONF.get('common')
-    versions_repo_url = config.get('packages_metadata_repo_url')
-    versions_repo_name = os.path.basename(os.path.splitext(versions_repo_url)[0])
-    versions_repo_target_path = os.path.join(
-        CONF.get('default').get('work_dir'), REPOSITORIES_DIR, versions_repo_name)
-    package_list = []
-    try:
-        package_list = [
-            package for package in os.listdir(versions_repo_target_path)
-            if os.path.isdir(os.path.join(versions_repo_target_path, package)) and
-            os.path.isfile(os.path.join(versions_repo_target_path, package,
-                                        "".join([package, ".yaml"])))
-        ]
-    except OSError:
-        LOG.error("No packages found in versions repository directory")
-        raise
-
-    return package_list
 
 
 class ConfigParser(object):
