+++ tools/build_package.py
@@ -15,9 +15,9 @@
 
 import logging
 
-from lib import config
 from lib import distro_utils
 from lib import build_manager
+from lib import packages_manager
 from lib.versions_repository import setup_versions_repository
 
 LOG = logging.getLogger(__name__)
@@ -26,7 +26,7 @@ LOG = logging.getLogger(__name__)
 def run(CONF):
     setup_versions_repository(CONF)
     packages_to_build = (CONF.get('common').get('packages') or
-                         config.discover_packages())
+                         packages_manager.discover_packages())
     distro = distro_utils.get_distro(
         CONF.get('common').get('distro_name'),
         CONF.get('common').get('distro_version'),
